Do not allow to choose just SSLv2Hello
Change SecurityProtocolsPanel and SSLConnector to validate that
SSLv2Hello is set in conjunction with at least one SSL/TLS version.
Otherwise it would lead to an exception when receiving/starting the
connections.
Change ConnectionParam and ProxyParam to validate that the SSL/TLS
versions persisted can be set/used, falling back to default versions if
not.