Clear cached certs when setting Root CA cert
Change CachedSslCertifificateServiceImpl to clear the certificates cache
when a new Root CA certificate is set, to ensure that it's used the new
Root CA cert for issuing the (new) certs. Also, sync when the Root CA
cert is set to avoid concurrency issues.
Change ExtensionDynSSL to set the cert through the previously mentioned
class, to ensure the cache is cleared.