Allow to configure the parse response limit
Change Spider to allow to configure maximum size in bytes for a response
to be parsed.
Add tests to assert the expected behaviour of DefaultParseFilter. Also,
normalise its behaviour (null/empty handling of the HTTP message),
tweak the case of constants and other minor tweaks.

Fix #1313 - Spider - Allow to configure the size limit of parseable
responses