Allow to set a tab scrollable in std dialogue
Change StandardFieldsDialog to allow to set a tab/panel scrollable, to
properly show all the fields (that is, by giving an indication with the
scrollbar that a field is hidden, partially or totally).
Change SpiderDialog to set the Advanced tab as scrollable (needed
because the tab is showing more options).