Show all messages sent by the spider
Change the Spider to show (and allow access to, through the API) all the
messages sent by the Spider, even if they were not processed further by
the Spider (for example, the spider was stopped after the message was
received or the message was filtered by the parser filters). Also, show
the reason why the messages were not processed.
Change ExtensionFactory to set the core ResourceBundle to the core
extensions so that they can use it without depending on the Constant
class (like non-core extensions can do).
Change I18N class to expose the core ResourceBundle to support the above
case.
Update tests to match the new behaviour, parser filters can now return
the reason why a message was filtered.

Fix #3392 - Show all messages sent by the Spider