Allow to change the Alerts options through the API
Change CoreAPI to add actions/views to allow to change/get the values of
the Alerts options.
Change ExtensionAlert to expose the method that reloads the file with
alert overrides (needed for the API action).
Change AlertParam to treat negative values as zero when setting the
maximum number of instances that should be included in the report.

Fix #3443 - Expose Alerts options through the ZAP API