Add Export URLs menus to Sites context menu
Add menu items "Export All URLs to File..." and "Export Selected URLs to
File..." to the context menu of Sites tree.
Refactor the Export URLs menus to require an Extension, it's a
dependency (failing to set it would lead to exceptions).

Fix #3446 - Add ability to export a Site Map via Context Menu