Use local proxy address for Browse API menu
Change ExtensionAPI to use the address of the local proxy for Browse API
menu, which does not require the browser to be proxying through ZAP
(previously it used always the URL "http://zap/").
Change API to expose the base URL (and change the API class variables to
be final, those do not need to be changed at runtime).

Fix #3461 - Browse API Doesn't work when browser isn't using ZAP as
proxy