Use another lock for add-on addition/removal
Change AddOnLoader to use other object (that the instance itself) to
synchronise the addition and removal of an add-on. The instance is
already being used to lock the loading of the classes, now shared with
the AddOnClassLoader, which increases the chances of a deadlock when an
add-on is being added/removed and a class being loaded by another thread
spawned by the affected add-on.

Related to #3494 - Fix class loading lock with scripts