Print args and error msg when failed to parse args
Change ZAP class to print the arguments and the exception's message when
failed to parse the arguments.
Change CommandLine to accept null arguments, remove print statements
(already done by ZAP class, when printing the exception's message) and
add a message to an exception thrown.
Add tests to CommandLineUnitTest to assert expected behaviour.