Increase CFU timeout when getting ZapVersions.xml
Change ExtensionAutoUpdate to give more time (up to 60s) to obtain the
ZapVersions.xml file, the current time (30s) might not be enough if the
1st request/URL times out and the 2nd takes more time than usual (in
case it requires proxy authentication, DNS queries take longer, or the
connection is slow/throttled).
Based on ZAP logs provided in IRC and mailing list, which indicated that
the CFU failed (because of the max time) but the download of the file
ended up happening.