Return correct error on missing auth parameters
Change AuthenticationAPI to use the existing method from ApiUtils to
obtain the context (reduce code duplication), also remove unused method.
Change ApiUtils to throw a "missing parameter" exception if the required
parameter is not provided and just throw "illegal parameter" if provided
but not valid (used by the method that obtains the Context, using the ID
provided in the parameters).
Add tests to assert the expected behaviour.