Show plugin as OFF, in policy panels, if disabled
Change CategoryTableModel to return OFF alert threshold is the plugin is
disabled, that represents better the state of the plugin (the enabled
state is not directly shown in the table).
Change AllCategoryTableModel to take into account the enabled state of
the plugins (assume OFF if disabled) when showing the alert threshold of
each category.