Refactor baseline and api scan scripts
The new zap_common.py script has an underscore in its name as you cant
import files with a minus in :( Dont really want to change the other
file names as that will break things for people.
I'm sure the refactoring could be better, but its a start ;)
Feel free to comment but please dont check in yet (in the unlikely event
they are approved first time;) - I'm running them against the Mozilla
sites we use the existing scripts with as a sanity check.