Do not traverse Sites tree while active scanning
Change HostProcess to collect all the messages that should be scanned
before starting the actual scan, instead of traversing the Sites tree
each time a node/message needs to be scanned, to prevent changes done to
the Sites tree from affecting the active scan (e.g. by new messages
added to the Sites tree).
Change HistoryReference to allow to keep the HttpMessage in memory when
constructing it (it's now immediately used when scanning).

Fix #3539 - Collect messages to scan before active scanning