Notify when a note is set to a message
Add the new event to HistoryReferenceEventPublisher.
Change HistoryReference to notify when a note is set to a message.
Change NotesAddDialog to remove the manual notification of the change,
it can now be consumed when listening for the new event, also, close
the dialogue itself (no longer requiring ExtensionHistory) and properly
clear the dialogue when disposing.
Change ExtensionHistory to handle the new event and deprecate/remove
code no longer used/needed.