Correct add-on version used in warning messages
Change AddOnRunIssuesUtils to use the semantic version (if/when)
declared in the manifest file in the warning messages about mismatch
version of the dependencies, the add-ons are not yet using semantic
version as the main version (which would lead to wrong version being
shown).
Change AddOn to expose the semantic version declared in the manifest
file (just for use in warning messages).