Update add-on files on newer versions
Change AddOnLoader to check and update the files of add-ons that have
been implicitly updated (e.g. newer version bundled in ZAP or manually
copied to plugin directory).
Change AddOnInstaller to allow to update the files of an add-on.

Fix #3771 - Add-on files not updated when running newer ZAP/core version