Fix unicode encoding errors when writing reports
Fix unicode encoding errors when writing reports in a way that supports
both Python 2.7 and Python 3.5. The string in Python 2.7 is a byte string,
and the default system encoding is ascii, which is why it breaks with just
mode w as it tries to encode a utf-8 string as ascii. Changing the mode to
wb fixes this as then it doesn't try to encode it further when writing. But,
in Python 3, the string is unicode so needs to be encoded as bytes before
being written with mode b.