Do not try to read plugin dir if it does not exist
Change AddOnCollection to return if the provided plugin directory does
not exist or if it's not a directory, instead of continuing and try to
read it (leading to an exception). Also, change the log level to WARN,
it's not a ZAP error/issue if it does not exist or it's not a directory.