Dynamically set the required SQL parameters
Change DbSQL to allow to specify the number of parameters that should be
used in a given query and remove the restriction of max parameters
allowed (now explicitly set by callers).
Change SqlTableHistory to specify the number of query parameters when
obtaining the messages by history type.
Change hsqldb.properties and mysql.properties to parameterise the
queries that allow to obtain the messages by history type.

Fix #3827 - mysql NULL value in where in CLAUSE