Reduce ScriptEngineFactory.getScriptEngine() calls
Change ScriptEngineWrapper to keep a reference to ScriptEngineFactory
and use it when obtaining data related to the engine/factory (mainly
extensions (and to obtain the engines).
Change ExtensionScript to use the methods provided by the wrapper to
obtain the factory and the extensions, calls to
ScriptEngineWrapper.getEngine().getFactory() are costly with the newer
Jython engine, which makes ZAP slower to start and open the New Script
dialogue.
Change DefaultEngineWrapper to no longer override getExtensions(), now
done by base class.