Add SNI support to Local Proxy
Change SSLConnector to rely on SNI if the domain is not known when
creating the SSL/TLS tunnel.
Change ProxyThread to check if the first message received is a SSL/TLS
handshake and start the SSL/TLS connection if so, also, tweak an
exception message to differentiate the case where the domain is unknown
(relying on SNI) from when it's known.

Fix #1015 - Support Server Name Indication