Normalise form-based login URL validation/usage
Change FormBasedAuthenticationMethodType to validate and use the login
URL always in encoded form, instead of encoded in some places and
decoded in others. Attempting to set a decoded/improperly encoded login
URL could lead to a URIException preventing the URL from being set.
While the URL had to be encoded to be set it would later be used as if
it was not encoded (leading to double encoding).
This change might break existing API users that are setting a login URL
partially/improperly encoded.

Related to #3854 - Error whilst using Api - ERROR
ExtensionUserManagement - Unable to persist Users