Refactor target access in docker scripts
Extract a function into zap_common.py that accesses the target URL and
tweak the error message (to be more clear that's ZAP that failed to
access the target).
Change zap-baseline.py and zap-full-scan.py to use the new function.

Related to #3902 - ZAP error messages when DNS is not working in docker
scripts could be more forth coming