Allow to unload ContextPanelUsersSelectComboBox
Change ContextPanelUsersSelectComboBox to allow to unload it, to be
GC'ed when no longer needed.
Change UsersListModel to allow to unload it (helper class used by the
previous class, this class adds itself to "core" table model to get the
latest users available, thus needs to be removed when no longer needed).