Do not allow to open a FindDialog with null parent
Change FindDialog to throw an exception if the parent is null, as
there's no parent to cache to.
Change PopupFindMenu to drop the use of parent JFrame, the text
component might not have one (e.g. contained in parentless JDialog).
Change ExtensionEdit to obtain the ancestor Window and use it as the
parent of the text component, it correctly handles the cases of a parent
JDialog and JFrame.