Do not attempt to parse empty Cookie headers
Change HttpRequestHeader to skip the parsing of empty Cookie headers
(there's nothing to parse).
Change the extensions that triggered the errors to handle the exception
thrown when the cookies are malformed and change the log level to
warn in the existing log statement (most likely not an error in ZAP).

Fix #3965 - Empty Cookie header causes errors