Allow to delete History entries with keyboard
Add an action to LogPanel (that is, the table that shows the history
entries) to allow to delete the history entries with keyboard shortcut.
Add a method to ViewDelegate to allow to obtain the default delete
keyboard shortcut (to be used/shared by all the UI components).
Implement the new method in View.
Move methods to delete/purge history entries from PopupMenuPurgeHistory
into ExtensionHistory and made the main method public, now used by the
pop up menu and LogPanel.
Set the default delete keyboard shortcut as accelerator of
PopupMenuPurgeHistory.

Fix #3626 - Allow to delete messages with keyboard shortcut