Differentiate GET/POST array parameters
Change VariantAbstractQuery to use indexed names for array parameters,
for example, if param[]=x&param[]=y use as names param[0] and param[1],
respectively.
Add tests to assert the expected behaviour of VariantAbstractQuery and
normalise some of its behaviour (that is, validate that valid parameters
are set, prevent modifications to internal parameter list, and do not
accumulate parameters on successive calls).

Fix #2496 - ZAP only report the first alert in array