Do not validate cookie path
Add ZapCookieSpec, same behaviour as CookieSpec already in use but
without the path validation. The path does not need to be validated (RFC
6265).
Change HttpSender to use the ZapCookieSpec as browser and default
CookieSpec.
Add some tests to assert the expected behaviour.

Fix #4079 - Cookie with unknown path is falsely rejected by ZAP