Normalise API authentication methods' responses
Change authentication methods to return the JSON API responses wrapped
in an object instead of returning the data directly, so that the Python
ZAP API client is able to correctly read the data (expects and removes
the wrapper object).

The data returned is now, for example:
```JSON
{
  "method": {
    "port": "443",
    "host": "example.com",
    "methodName": "httpAuthentication",
    "realm": "example"
  }
}
```

instead of:
```JSON
{
  "port": "443",
  "host": "example.com",
  "methodName": "httpAuthentication",
  "realm": "example"
}
```

This change breaks existing custom clients that consume the JSON format.
The structure of XML format is unchanged for all methods except "manual"
(normalised with the other methods).

Fix #4097 - python api returns the wrong content for
authentication.get_authentication_method()