Return API alerts summary in a JSON object
Change CoreAPI to return the JSON API response of alerts summary wrapped
in an object instead of returning the data directly, so that the Python
ZAP API client is able to correctly read the data (expects and removes
the wrapper object).

The data returned is now, for example:
```JSON
{
  "alertsSummary": {
    "High": 0,
    "Low": 3,
    "Medium": 1,
    "Informational": 1
  }
}
```

instead of:
```JSON
{
  "High": 0,
  "Low": 3,
  "Medium": 1,
  "Informational": 1
}
```

This change breaks existing custom clients that consume the JSON format.

Fix #4111 - ZAP Python API: zap.context.alerts_summary() returns 0