Remove filter extension from core functionality
Change CoreFunctionality to not add the ExtensionFilter, superseded by
scripts and Replacer add-on. The filter functionality is deprecated (UI
message) since version 2.4.0.
Add deprecated annotation and JavaDoc tag to filter related classes, for
ExtensionFilter also override the deprecated method. The remove of the
code and related resources should be done in a following release, once
all known add-ons are no longer using it.