Refactor redirection validation in active scans
Change AbtractPlugin and Analyser to use HostProcess to validate the
redirections, to reduce code duplication.
Change HostProcess to provide a HttpRedirectionValidator and a
HttpRequestConfig for validation of redirects.
Change HttpRedirectionValidator to have notifyMessageReceived as a
default method, the method is provided for convenience and not all
implementation require it, allowing the interface to be functional.
Change DefaultHttpRedirectionValidator to not implement the now default
method.