Warn and prevent snapshot to current session file
Change MenuFileControl to warn and prevent the snapshot (and save) of
the session to current session file in case the current session file is
accidently selected (which would fail with a NullPointerException).
Change WritableFileChooser to allow to show an error dialogue, used by
MenuFileControl's custom WritableFileChooser.
Add new error message/title to Messages.properties.