Handle null value in ApiResponseElement
Change ApiResponseElement to handle the null value when converting to
XML and JSON, otherwise it would lead to NullPointerException in API
when returning the response. For JSON convert to null value and for XML
to an empty element.
Add tests to assert the expected behaviour with null and non-null value.