Allow to check if several techs are enabled
Change AbstractPlugin to allow to check if any of several techs are
enabled for the scan, to make it more straightforward when checking
several techs for the same test/scan.
Change TechSet to allow to check if any of several techs are included.
Tweak JavaDoc in Plugin.inScope to also mention tests.