Allow to get the current parameter in IV scripts
Change VariantCustom to allow Input Vector scripts to get the current
parameter being tested so that they can reliably know which parameter is
being tested (instead using custom names or the name being injected,
which might not be the same as the current parameter).
Change Input Vector default template.js to make use of the new method.