Normalise Source instantiations
Change ExtensionAntiCSRF, PassiveScanThread, and HtmlContextAnalyser to
only use the response body when instantiating the Source, the header is
not necessary, the body is already using the expected charset (as
specified in the header or determined from inspecting the body).
Other classes (spider parsers and corresponding tests) were already
using just the body.