Use same help page in request editor dialogues
Change ExtensionHistory to use the same help page for both request
editor dialogues (that is, use "ui.dialogs.manreq" in resend one), both
have the same behaviour/functionality, the only difference is that one
starts with a default request while the other with an existing one.

Related to zaproxy/zap-core-help#175.