Use FileNameExtensionFilter where possible
Change to use FileNameExtensionFilter instead of custom FileFilter where
just the file extension of the file was being checked. For the other
cases correct lower case comparisons to use a neutral locale.
Tweak FilenameExtensionFilter to only do lower case conversion of the
extension once, add JavaDoc, and license header.

Related to #4327 - ZAP relies on default locale when it shouldn't