Do not throw MissingResourceException in I18N
Change I18N to not throw a MissingResourceException if the message does
not exist when obtaining it using getString(String), to avoid failing
hard when the message does not exist, while the cases are rare they
happen and it's preferable to show something than nothing at all. Also,
the method getString(String, Object...) was already catching the
exception (and returning the key itself). Now both methods will log an
error indicating that the key is missing and return the key itself.
Change View and WebUI to no longer rely on the exception, instead check
(explicitly) if the message exists.