Fix main layouts
Change WorkbenchPanel to hide the panels instead of relying on the
behaviour of the UI components (which removes the panel from the current
parent before adding it to itself), as that would end up removing them
from the internal list of panels of the TabbedPanel2 (per changes done
in #4511).
Change TabbedPanel2 to allow to hide all the tabs and add a panel hidden
(adds to the internal list without displaying the tab), to support the
above case.