Keep panel of selected context selected
Change SiteMapPanel to not change the selected panel (to main context
panel) if the panel belongs to the selected context, it's more likely
that the user wants to keep using the previously selected panel than go
to main panel every time the context is double clicked.
Change AbstractParamContainerPanel and AbstractParamDialog to allow to
check if a panel or its child panels is selected, to support the above
use case.