Fix Issue 3447 - Params Table Truncation
When DatabaseException is caught check if the cause message contains
"truncation". If so log an warn message (if debug is enabled log the
stack trace). If the cause does not contain "truncation" log an error as
per the existing behavior.

Since the values field (column) is ~8MB it should be rare that this
occurs. (In fact to test/recreate I had to temporarily edit the DB
.script file to make the values column ~80KB and reload a page with a
~10KB header value a bunch of times.) While the user may want to be
aware of what's happening it is unlikely to impact their use of ZAP for
"normal" use cases.

Fixes zaproxy/zaproxy#3447