Add JSON-based authentication method
Add an authentication method, JsonBasedAuthenticationMethodType, that
allows to send a JSON object in the request body. It can be configured
through the API with the name "jsonBasedAuthentication".
Extract a class, PostBasedAuthenticationMethodType, to be used by
FormBasedAuthenticationMethodType and the new auth method, the behaviour
is exactly the same except for the data they handle (application/json or
application/x-www-form-urlencoded) and the POST data required.
Add libraries used to quote the authentication credentials when injected
into the request body.
Add JSON Content-Type to HttpHeader.

Fix #2439 - Configure Authentication with Json object