Track the directories with scripts
Change ExtensionScript to add the scripts of the directories also when a
script engine or type is added (they were already being removed). This
ensures that the scripts of the directories are added when an add-on
with script engines or types is installed, updated, or finished
loading (the case of Python engine added asynchronously).