Load/use test resources with classpath's path
Change tests to load/use the test resources using the classpath's path,
which does not depend where the resources are located in the filesystem.
Change build.xml file to add the test resources to the classpath (copied
previously to buildtest dir).
Change Eclipse's classpath to include the test resources.
Change WithConfigsTest to use a temporary directory for ZAP home and
installation directory (and remove files no longer needed).
Add helper test class to load the resources.