Ignore null script engine when listing engines
Change ScriptEngineWrapper to allow to tell if the engine should be
visible. Also, add workaround to have NullScriptEngineWrapper as not
visible, it's not a functional engine and it shouldn't be visible (it's
already hidden in the UI but was returned through the API as " : Null").
Change ExtensionScript to not include engines that are not visible when
listing engines.