Provide details of why the add-on is not valid
Change ExtensionAutoUpdate to provide details of why an add-on is not
valid when installing a local add-on, before it would just say that the
add-on is not valid, now it will show the actual reason why it is not
valid.
Change AddOn to provide the result of the validation to allow callers to
show a more appropriate message, also, avoid reading the manifest twice
by using the manifest already read when validating.