Allow add-ons to declare a resource bundle
Add-ons can now declare a resource bundle through the manifest file:
```XML
<bundle prefix="example">com.example.Messages</bundle>
```
the prefix is used to add the ResourceBundle into the I18N instance in
the Constant class. The bundle is also accessible through the add-on
itself.

The extensions of the add-on will use the declared bundle if none is
found in the default location.

Fix #3841 - Allow add-ons without extensions to have resource messages