Stop relying on ClassLoaderUtil
Deprecate ClassLoaderUtil and stop using it throughout the codebase. The
class expects a URLClassLoader as system class loader which is no longer
the case with newer Java (9+) versions (the class was already a no-op in
those cases and others like Webswing that also uses a different system
class loader).
Obviously, this change will break applications/configurations relying on
that behaviour, those applications need to be updated to include the
required resources in the classpath (e.g. Messages.properties file under
the lang dir and libraries under the lib dir).

Related to #2602 - Java 9