Correct install dir detection from ZAP JAR
Change Constant class to convert the code source location to a file
system path and check if it's a directory, if not (i.e. a JAR) get the
parent directory. Otherwise it would use the JAR as the installation
dir, failing to start or leading to errors as it attempts to read the
files with the JAR in the path (e.g. "/dir/zap.jar/db/zapdb.script").