Bundle Messages.properties and vulnerabilities.xml
Change build.xml to include the files Messages.properties and
vulnerabilities.xml in the zap.jar.
Change I18N and Vulnerabilities to fallback to the bundled files if the
corresponding files are not in the file system.
Tweak LocaleUtils to not attempt to read the lang directory (for
Messages.properties) as it might no longer exist and if it exists log a
warn (instead of error) if no files are found (it would use the bundled
file).
Change VulnerabilitiesLoader to allow to load a vulnerabilities.xml file
using an InputStream (e.g. the bundled one) and return an empty list
of vulnerabilities if the lang directory is not found (to fallback to
bundled file).
Change ZapXmlConfiguration to allow to read a configuration from an
InputStream.

Part of #4771 - Bundle required files in the zap.jar