Adjust log msg/levels related to plugin dirs
Change AddOnCollection to provide a more informative message when the
plugin directory does not exist.
Change AddOnLoader to use debug and warn levels instead of error when
the directory does not exist and if it's not a directory, respectively,
this only applies to loading of JARs (which are not required for ZAP
to start/work).

Related to #4771 (i.e. when running ZAP with just the JAR+libs).