Remove JavaFX runtime library
The library is not used by core. Only one add-on (Browser View) is using
it, moreover the add-on already copes with the missing library and when
missing it informs the user that JavaFX is required.
Remove the library from LEGALNOTICE.md.
Update the classpath and remove the library.

Part of #4780 - Tidy up dependencies