Properly re-parent panels in TabbedPanel2
Change TabbedPanel2 to set itself as parent also to non-visible panels
when the visible panels are set to ensure those panels have the correct
parent (not just the visible ones). Also, don't re-parent if the panel
being added is hidden, as it means it's being shown (thus child) of
other TabbedPanel2.

Fix #4789 - Break tab is not automatically shown in full layout when
breakpoint is hit