Allow to enable dev mode through the command line
Add a command line option, `-dev`, to enable "dev mode", that is, enable
development related utilities/functionalities. For example, show the
error counter in the footer tool bar and have the license implicitly
accepted (thus not requiring to show/accept the license each time a
new home is used). This allows to use those functionalities even if
using a weekly or main release, when developing add-ons.