Allow consumers to remove themselves during events
Change SimpleEventBus to allow consumers to remove themselves during
event notification, which before would lead to a deadlock.
The notification is no longer done under a lock and can be done
concurrently (any modification on the list of consumers leads to a copy
of the list, CopyOnWriteArrayList, preventing the above deadlock and
ConcurrentModificationExceptions).
Add test to assert the expected behaviour.