Add Java 9 to Travis CI builds
Change .travis.yml to also run the build with Java 9.
Change tests to not use PowerMock which has issues with Java 9, moreover
most of the tests didn't actually require it (e.g. Constant.messages is
accessible).
Change PluginFactory to allow to initialise it without using add-ons (to
ease the test setup).