Address compilation warnings with Java 10
Address compilation warning with usage of ToolKit.getMenuShortcutKeyMask
by suppressing it (the new method is only available in Java 10).
Add a getter to ViewDelegate for KeyStroke with menu shortcut key mask,
to reduce the occurrences of the above method (and slightly reduce code
duplication).
Update codebase to use the new method.
Change .travis.yml to also run the build with Java 10.