Test initialisation of home directory
Add tests to assert that the expected files/directories are created in
the home directory.
Change Constant to replace the malformed config file also when copying
from bundled file and to use the name of the plugin directory instead of
the previous path (to ensure it's created always the expected path).