Log duplicated default accelerators just once
Change ExtensionKeyboard to ignore menu items that were already logged,
for example, it would log again if an add-on was installed (since the
panel menus are recreated). Also, use the identifiers instead of the
menus to not prevent the menus from being GC (and the corresponding
add-ons from being unloaded).
Extract some methods to reduce code duplication and make it easier to
read.

Related to #4874 - Ignore (and log) duplicated default accelerators