Fix CME in passive scanner
Change PassiveScannerList to use a CopyOnWriteArrayList which allows
modifications while the list is being iterated, the latter happens very
frequently (each time a message is persisted) the former less often
(when add-ons with scanners are installed/updated or scan tags changed).
Add tests to assert the expected behaviour.

Fix #4911 - ConcurrentModificationException thrown in PassiveScanThread