Use default user agent in all default messages
Change ConnectionParam to set the default user agent into the
HttpRequestHeader to be used in the default request headers, instead of
requiring users to pass a ConnectionParam, the latter is error prone.
Change HttpRequestHeader to allow to set (and use) the default user
agent and deprecate the constructor that accepts a ConnectionParam.
Change codebase to no longer use the (now) deprecated constructor.

Fix #4846- User-Agent changes from configured value