Don't parse (plain) API parameters as JSON
Change API to use the utility class to process the API parameters before
putting them in the JSONObject, to avoid parsing them as JSON (for the
API they are plain text data, it's up to the actual endpoints to parse
them if needed).
Add some JavaDoc to JsonUtil and tweak the quoting method to use the
same logic as the AbstractJSON._processValue, faster than parsing the
whole value.
Add tests to assert the expected behaviour.