Remove redundant add-on validations
Remove validations done before creating add-ons since the creation of
the add-on already validates it.
Add a convenience method to AddOn that attempts to create the add-on and
change AddOnCollection and ExtensionAutoUpdate to use it, while removing
the redundant add-on validations done beforehand (for the latter class
keep the file name validation to differentiate from ZAP packages).