Correct SSE Content-Type checks
Change HttpMessage/ZapGetMethod to check if the Content-Type contains
the SSE type/subtype instead of checking if it's equal, to allow
parameters in the media type.
Expose HttpHeader methods around Content-Type handling, now used in
prior check.
Add tests to assert the expected behaviour of HttpMessage.

Fix #5003 - SSE are not supported in some cases