Do not disable break buttons if msgs still caught
Change BreakPanelToolbarFactory to keep track of how many messages are
currently caught to just disable the buttons when none is left. Now the
buttons are no longer disabled after stepping/dropping only when
continuing (which forwards everything) or no messages left.
Remove the boolean parameter from setStep/setDrop and rename to step and
drop, respectively, the parameter is no longer needed the break buttons
are no longer affected (and the parameter was always the same, true).
Update BreakPanel to reflect the changes in the step/drop methods.

Fix #4623 - Message caught in Break tab but break buttons remain
disabled