Correct break events with multiple messages
Change BreakpointMessageHandler2 to only send the active/inactive events
when the message is actually active/inactive not when a breakpoint is
hit, otherwise it would mislead the consumers by sending multiple active
events (when only one message is active).
Add a new event to BreakEventPublisher for when a breakpoint is hit.