Use varargs on registrations in EventBus
Change EventBus to use vararg for events when registering publishers and
consumers, to make the registrations less verbose (i.e. don't require
the explicit creation of an array).
Change existing registrations to not create the array.
Change ScanEventPublisher to not expose the array of events to avoid
accidental changes to its contents, update usages accordingly.