Do not remove users when changing auth method
Change AuthenticationMethodType to expose the type of the credentials
it creates to allow to check if existing users' credentials need to be
reset, also, remove API related code that removed the users.
Update AuthenticationMethodType implementations for the above changes.
Change AuthenticationAPI and ContextAuthenticationPanel to no longer
remove the existing users when changing the authentication methods,
instead reset their credentials if different type.

Fix #1067 - User conversion between changes of AuthenticationMethods