Do not show main release notes for daily release
Change ManageAddOnsDialog to not show the release notes button for daily
release, the release notes shown belong to main release not daily. Also,
tweak the message that shows the latest version to have a space before
the version.
Extract a method to reduce code duplication when adding the components
of the newer version.