Show add-on details in a panel instead of tool tip
Change Manage Add-ons dialogue to show the add-on details in a bottom
panel in the tabs Installed and Marketplace, to give more space and
properly render the changes of the add-ons (e.g. word wrap, links).
Change ZapLabel to disable caret updates instead of resetting the
position after setting the text to avoid scroll jumps.
Change AddOn to read the info URL also when creating the add-on from a
file.