Remove redundant options from default conf file
Remove all options that are already defined in the code to avoid having
to maintain the keys and values in two places.
Change default values in the code to match the ones used in the config
file.
Update installer to insert the options instead of replacing since the
start element is no longer present in the file.

Fix #5175 - ConversionException on weekly startup