Ensure The Scan Policy Sticks Between Active Scan Runs
I find this handy, hopefully others do to. Unfortunately as discussed in
the ticket there are still other changes needed on other panels for
consistency. (Even with this change the "Cancel" behavior is not
intuitive, however, I do feel this is more user friendly as-is when
running multiple scans.)

See zaproxy/zaproxy#3538 for further details.