Add tasks to create the distributions
Move distribution files to `zap` project:
 - Common files to `src/main/dist/`, also, replace hardcoded JAR names
 in `zap.sh` and `zap.bat` with tokens (replaced with actual name when
 building the distributions);
 - macOS files to `src/main/macOS/`, also, tweak tokens in `Info.plist`
 to use the default characters (`@` instead of `_`);
 - Debian files to `src/main/debian/`.

Remove Debian `control` file, now generated by the build.
Remove macOS build files `.gitignore` and `ZAP.icns` no longer needed,
the directory for the JRE is created as needed and the `icns` file is
copied when needed as well.
Tweak `Readme.txt` in the `plugin` directory to not refer to the
repository, not relevant when in the distributions, instead mention that
in `BUILDING.md`.

Add the following tasks to build the distributions:
 - `:zap:distCrossplatform`, to build the cross platform package (done
 manually before);
 - `:zap:distCore`, to build the core distribution, former Ant task
 `package-lite`;
 - `:zap:distDaily`, to build the daily distribution, former Ant task
 `package-daily`;
 - `:zap:distDebian`, to build the deb package, former Ant task
 `package-linux` (built Linux and deb package at the same time);
 - `:zap:distLinux`, to build the Linux distribution, former Ant task
 `package-linux`;
 - `:zap:distMac`, to build the dmg file, former Ant task
 `package-macosx`.

The distributions are built into `build/distributions/`.
Update `BUILDING.md` to mention the distribution tasks.

Signed-off-by: thc202 <thc202@gmail.com>